+++ demisto_sdk/commands/find_dependencies/tests/find_dependencies_test.py
@@ -3,8 +3,8 @@ import os
 
 import pytest
 from demisto_sdk.commands.common.git_tools import git_path
-from demisto_sdk.commands.find_dependencies.find_dependencies import \
-    PackDependencies
+from demisto_sdk.commands.find_dependencies.find_dependencies import (
+    PackDependencies, VerboseFile)
 from TestSuite.utils import IsEqualFunctions
 
 
@@ -121,7 +121,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -144,7 +147,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -165,7 +171,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -186,7 +195,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -216,7 +228,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -235,7 +250,10 @@ class TestDependsOnScriptAndIntegration:
             }
         ]
 
-        dependencies_set = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
+        dependencies_set = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input,
+                                                                          id_set=id_set,
+                                                                          verbose_file=VerboseFile(),
+                                                                          )
 
         assert len(dependencies_set) > 0
 
@@ -271,7 +289,10 @@ class TestDependsOnPlaybook:
             }
         ]
 
-        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set,
+                                                                        verbose_file=VerboseFile(),
+                                                                        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -302,7 +323,10 @@ class TestDependsOnPlaybook:
             }
         ]
 
-        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set,
+                                                                        verbose_file=VerboseFile(),
+                                                                        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -333,7 +357,10 @@ class TestDependsOnPlaybook:
             }
         ]
 
-        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set,
+                                                                        verbose_file=VerboseFile(),
+                                                                        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -360,7 +387,10 @@ class TestDependsOnPlaybook:
                 }
             }
         ]
-        found_result_set = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
+        found_result_set = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                            id_set=id_set,
+                                                                            verbose_file=VerboseFile(),
+                                                                            )
 
         assert len(found_result_set) == 1
         found_result = found_result_set.pop()
@@ -390,7 +420,10 @@ class TestDependsOnPlaybook:
             }
         ]
 
-        found_result_set = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
+        found_result_set = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                            id_set=id_set,
+                                                                            verbose_file=VerboseFile(),
+                                                                            )
 
         assert len(found_result_set) > 0
 
@@ -424,7 +457,71 @@ class TestDependsOnPlaybook:
         ]
 
         found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
-                                                                        id_set=id_set)
+                                                                        id_set=id_set,
+                                                                        verbose_file=VerboseFile(),
+                                                                        )
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+    def test_collect_playbooks_dependencies_skip_unavailable(self, id_set):
+        """
+        Given
+            - A playbook entry in the id_set.
+            -
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the playbook depends on.
+        """
+        expected_result = {
+            # playbooks:
+            ('Slack', False), ('Indeni', True),
+            # integrations:
+            ('FeedAlienVault', False), ('ipinfo', True), ('FeedAutofocus', True),
+            # scripts:
+            ('GetServerURL', False), ('HelloWorld', True),
+        }
+        test_input = [
+            {
+                'Dummy Playbook': {
+                    'name': 'Dummy Playbook',
+                    'file_path': 'dummy_path',
+                    'fromversion': 'dummy_version',
+                    'implementing_scripts': [
+                        'GetServerURL',
+                        'HelloWorldScript',
+                    ],
+                    'implementing_playbooks': [
+                        'Failed Login Playbook - Slack v2',
+                        'Indeni Demo',
+                    ],
+                    'command_to_integration': {
+                        'alienvault-get-indicators': '',
+                        'ip': 'ipinfo',
+                        'autofocus-get-indicators': '',
+                    },
+                    'tests': [
+                        'dummy_playbook'
+                    ],
+                    'pack': 'dummy_pack',
+                    'incident_fields': [
+                    ],
+                    'skippable_tasks': [
+                        'Print',
+                        'Failed Login Playbook - Slack v2',
+                        'alienvault-get-indicators',
+                        'GetServerURL',
+                    ]
+                }
+            },
+        ]
+
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set,
+                                                                        verbose_file=VerboseFile(),
+                                                                        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -441,7 +538,8 @@ class TestDependsOnLayout:
         Then
             - Extracting the packs that the layout depends on.
         """
-        expected_result = {("FeedMitreAttack", True), ("PrismaCloudCompute", True), ("CommonTypes", True), ("CrisisManagement", True)}
+        expected_result = {("FeedMitreAttack", True), ("PrismaCloudCompute", True), ("CommonTypes", True),
+                           ("CrisisManagement", True)}
 
         test_input = [
             {
@@ -463,7 +561,10 @@ class TestDependsOnLayout:
             }
         ]
 
-        found_result = PackDependencies._collect_layouts_dependencies(pack_layouts=test_input, id_set=id_set)
+        found_result = PackDependencies._collect_layouts_dependencies(pack_layouts=test_input,
+                                                                      id_set=id_set,
+                                                                      verbose_file=VerboseFile(),
+                                                                      )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -480,7 +581,12 @@ class TestDependsOnIncidentField:
         Then
             - Extracting the packs that the incident field depends on.
         """
-        expected_result = {("Expanse", True), ("IllusiveNetworks", True), ("Carbon_Black_Enterprise_Response", True), ("Phishing", True)}
+        expected_result = {
+            # incident types
+            # ("Expanse", True), ("IllusiveNetworks", True),
+            # scripts
+            ("Carbon_Black_Enterprise_Response", True), ("Phishing", True)
+        }
 
         test_input = [
             {
@@ -501,7 +607,10 @@ class TestDependsOnIncidentField:
         ]
 
         found_result = PackDependencies._collect_incidents_fields_dependencies(
-            pack_incidents_fields=test_input, id_set=id_set)
+            pack_incidents_fields=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -545,7 +654,10 @@ class TestDependsOnIndicatorType:
         ]
 
         found_result = PackDependencies._collect_indicators_types_dependencies(
-            pack_indicators_types=test_input, id_set=id_set)
+            pack_indicators_types=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -560,7 +672,8 @@ class TestDependsOnIntegrations:
         Then
             - Extracting the packs that the integration depends on.
         """
-        expected_result = {("HelloWorld", True), ("Claroty", True), ("EWS", True), ("CrisisManagement", True), ("CommonTypes", True)}
+        expected_result = {("HelloWorld", True), ("Claroty", True), ("EWS", True), ("CrisisManagement", True),
+                           ("CommonTypes", True)}
 
         test_input = [
             {
@@ -580,7 +693,10 @@ class TestDependsOnIntegrations:
         ]
 
         found_result = PackDependencies._collect_integrations_dependencies(
-            pack_integrations=test_input, id_set=id_set)
+            pack_integrations=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -610,7 +726,11 @@ class TestDependsOnIncidentType:
         ]
 
         found_result = PackDependencies._collect_incidents_types_dependencies(
-            pack_incidents_types=test_input, id_set=id_set)
+            pack_incidents_types=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -643,7 +763,10 @@ class TestDependsOnClassifiers:
         ]
 
         found_result = PackDependencies._collect_classifiers_dependencies(
-            pack_classifiers=test_input, id_set=id_set)
+            pack_classifiers=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -658,7 +781,8 @@ class TestDependsOnMappers:
         Then
             - Extracting the packs that the mapper depends on.
         """
-        expected_result = {("AccessInvestigation", True), ("CommonTypes", True), ("PrismaCloud", True), ("BruteForce", True)}
+        expected_result = {("AccessInvestigation", True), ("CommonTypes", True), ("PrismaCloud", True),
+                           ("BruteForce", True)}
 
         test_input = [
             {
@@ -680,7 +804,10 @@ class TestDependsOnMappers:
         ]
 
         found_result = PackDependencies._collect_mappers_dependencies(
-            pack_mappers=test_input, id_set=id_set)
+            pack_mappers=test_input,
+            id_set=id_set,
+            verbose_file=VerboseFile(),
+        )
 
         assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
@@ -706,9 +833,35 @@ def test_search_packs_by_items_names_or_ids(item_names, section_name, expected_r
 class TestDependencyGraph:
     def test_build_dependency_graph(self, id_set):
         pack_name = "ImpossibleTraveler"
-        found_graph = PackDependencies.build_dependency_graph(pack_id=pack_name, id_set=id_set)
+        found_graph = PackDependencies.build_dependency_graph(pack_id=pack_name,
+                                                              id_set=id_set,
+                                                              verbose_file=VerboseFile(),
+                                                              )
         root_of_graph = [n for n in found_graph.nodes if found_graph.in_degree(n) == 0][0]
         pack_dependencies = [n for n in found_graph.nodes if found_graph.in_degree(n) > 0]
 
         assert root_of_graph == pack_name
         assert len(pack_dependencies) > 0
+
+    def test_build_dependency_graph_include_ignored_content(self, id_set):
+        """
+        Given
+            - A pack name which depends on unsupported content.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the pack dependencies with unsupported content.
+        """
+
+        pack_name = "ImpossibleTraveler"
+        found_graph = PackDependencies.build_dependency_graph(pack_id=pack_name,
+                                                              id_set=id_set,
+                                                              verbose_file=VerboseFile(),
+                                                              exclude_ignored_dependencies=False
+                                                              )
+        root_of_graph = [n for n in found_graph.nodes if found_graph.in_degree(n) == 0][0]
+        pack_dependencies = [n for n in found_graph.nodes if found_graph.in_degree(n) > 0]
+
+        assert root_of_graph == pack_name
+        assert len(pack_dependencies) > 0
+        assert 'NonSupported' in pack_dependencies